+++ inst/testme/_prologue/030.imports.R
@@ -1,0 +1,8 @@
+## Private future functions
+evaluate_expr <- future.tests:::evaluate_expr
+
+pcat <- future.tests:::pcat
+perase <- future.tests:::perase
+preset <- future.tests:::preset
+
+hpaste <- future.tests:::hpaste