+++ inst/testme/_prologue/099.future-setup.R
@@ -1,0 +1,6 @@
+## Use sequential futures by default
+oplan <- local({
+  oopts <- options(future.debug = FALSE)
+  on.exit(options(oopts))
+  future::plan(future::sequential)
+})